// packages/tokens/scripts/shared/dtcg/resolver/token/types/base/dimension/value/dimension-design-tokens-collection-token-value.ts
import type { DimensionDesignTokenValueUnit } from '../../../../../../design-token/token/types/base/types/dimension/value/members/unit/dimension-design-token-value-unit.ts';

export interface DimensionDesignTokensCollectionTokenValue {
  readonly value: number;
  readonly unit: DimensionDesignTokenValueUnit;
}